Fixed node test switching.
When a test is selected for a node, the sat (left) branch could
be associated with the negative label (unsat). If that is the case,
we now switch the branches and change the direction of the
inequality in the test. As a consequence, we now only consider one
direction for each primitive.